Refactor: Modularize test utilities
